feat: Pro版本更新4.5.8，支持CVE2026_43499漏洞（支持：高通+天玑，品牌：一加\OPPO\小米\红魔\联想Y700四代，内核范围：Linux 6.6.0~6.6.140、6.12.0~6.12.86）。

+++ Pro(众测开放中)/src/PermissionManager/app/src/main/java/com/linux/permissionmanager/fragment/HomeFragment.java
@@ -3,11 +3,10 @@ package com.linux.permissionmanager.fragment;
 import static com.linux.permissionmanager.AppSettings.HOTLOAD_SHELL_PATH;
 
 import android.app.Activity;
-import android.graphics.Color;
 import android.os.Bundle;
 import android.os.Handler;
+import android.os.Looper;
 import android.os.Message;
-import android.text.TextUtils;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -24,6 +23,7 @@ import androidx.fragment.app.Fragment;
 import com.linux.permissionmanager.AppSettings;
 import com.linux.permissionmanager.R;
 import com.linux.permissionmanager.bridge.NativeBridge;
+import com.linux.permissionmanager.model.HotloadMethod;
 import com.linux.permissionmanager.utils.ClipboardUtils;
 import com.linux.permissionmanager.utils.DialogUtils;
 
@@ -33,7 +33,7 @@ public class HomeFragment extends Fragment implements View.OnClickListener {
     private Activity mActivity;
     private String mRootKey = "";
     private boolean mIsHotload = false;
-    private String mHotloadMethod = "";
+    private HotloadMethod mHotloadMethod = null;
     private String lastInputCmd = "id";
     private String lastInputRootExecPath = "";
 
@@ -42,7 +42,7 @@ public class HomeFragment extends Fragment implements View.OnClickListener {
     private TextView tvAdbStatus;
     private TextView tvOneplusBypassStatus;
     private EditText mConsoleEdit;
-    public HomeFragment(Activity activity, String rootKey, boolean isHotload, String hotloadMethod) {
+    public HomeFragment(Activity activity, String rootKey, boolean isHotload, HotloadMethod hotloadMethod) {
         mActivity = activity;
         mRootKey = rootKey;
         mIsHotload = isHotload;
@@ -187,37 +187,22 @@ public class HomeFragment extends Fragment implements View.OnClickListener {
         DialogUtils.showInputDlg(mActivity, lastInputCmd, "请输入ROOT命令", null, inputCallback, null);
     }
 
-    enum InstallMode {
-        BOOT("Boot"),
-        HOTLOAD_REBOOT("HotLoadReboot"),
-        HOTLOAD_NO_REBOOT("HotLoadNoReboot");
-        private final String nativeValue;
-        InstallMode(String nativeValue) { this.nativeValue = nativeValue; }
-        public String getNativeValue() { return nativeValue; }
-    }
-
     private void onClickInstallSkrootEnvBtn() {
-        String insMode = InstallMode.BOOT.getNativeValue();
-        boolean isCve2026_43499 = TextUtils.equals(mHotloadMethod, "CVE-2026-43499");
-        if(mIsHotload) {
-            insMode = isCve2026_43499 ? InstallMode.HOTLOAD_NO_REBOOT.getNativeValue() : InstallMode.HOTLOAD_REBOOT.getNativeValue();
-        }
-        String err = NativeBridge.installSkrootEnv(mRootKey, insMode);
+        String err = NativeBridge.installSkrootEnv(mRootKey, mIsHotload ? "HotLoad" : "Boot",
+                mHotloadMethod != null ? mHotloadMethod.getConfigValue() : "");
         appendConsoleMsg(err);
-        if(mIsHotload && err.indexOf("OK") != -1) {
-            NativeBridge.runRootCmd(mRootKey, "rm -f " + HOTLOAD_SHELL_PATH);
-
-            if(isCve2026_43499) {
-                DialogUtils.showCustomDialog(mActivity, "确认", "请问要软重启吗？\n\n提示：当前环境已安装完成！模块也已激活。一般不需要软重启，当使用特殊模块（如改机型）才需要进行软重启。",null,
-                "不需要", (dialog, which) -> {
-                        dialog.dismiss();
-                        showSkrootStatus();
-                    }, "软重启", (dialog, which) -> {
-                        dialog.dismiss();
-                        NativeBridge.restartZygote64(mRootKey);
-                    }
-                );
-            }
+        if(mIsHotload && err.indexOf("OK") != -1) NativeBridge.runRootCmd(mRootKey, "rm -f " + HOTLOAD_SHELL_PATH);
+        if(mHotloadMethod == HotloadMethod.CVE_2026_43499) {
+            DialogUtils.showCustomDialog(mActivity, "确认", "请问要软重启吗？\n\n提示：当前环境已安装完成！模块也已激活。一般不需要软重启，当使用特殊模块（如改机型）才需要进行软重启。",null,
+            "不需要", (dialog, which) -> {
+                    dialog.dismiss();
+                    appendConsoleMsg("启动中，请稍后");
+                    new Handler(Looper.getMainLooper()).postDelayed(() -> showSkrootStatus(), 2000);
+                }, "软重启", (dialog, which) -> {
+                    dialog.dismiss();
+                    NativeBridge.restartZygote64(mRootKey);
+                }
+            );
         }
     }
